Change gex_EP_RegisterHandlers numentries to size_t
Using an unsigned type for the table size eliminates the possibility
of calls with a meaningless negative value for numentries.

Also add additional argument checking to disqualify invalid calls
with a useful error message, instead of a fatal signal.

+++ gasnet_internal.h
@@ -737,7 +737,7 @@ extern int gasneti_amregister( gex_AM_Entry_t *output,
                                int lowlimit, int highlimit,
                                int dontcare, int *numregistered);
 extern int gasneti_amregister_client(gex_AM_Entry_t *output,
-                                     gex_AM_Entry_t *input, int numentries);
+                                     gex_AM_Entry_t *input, size_t numentries);
 extern int gasneti_amregister_legacy(gex_AM_Entry_t *output,
                                      gasnet_handlerentry_t *input, int numentries);
